refactor(phase3): extract the REPL layer out of the cli.py god module (#182)
cli.py was ~1520 lines mixing the thin Click command layer with the entire
interactive REPL (a 223-line loop + ~13 slash-command/handler helpers + the
briefing/agent builders). The interactive layer is now its own module,
qracer/repl.py, leaving cli.py as the command layer.

- Moved to qracer/repl.py (verbatim): `_repl_loop`, all slash handlers
  (`_handle_*`/`_show_*`/`_report_agent_results`/`_split_action_schedule`), and the
  `_openrouter_provider` / `_build_briefing_composer` builders, plus BANNER/_HELP_TEXT.
- cli.py imports the four helpers its commands still call (`_repl_loop`,
  `_report_agent_results`, `_openrouter_provider`, `_build_briefing_composer`) from
  qracer.repl. One-directional (cli → repl → bootstrap), no import cycle.
- `install`/`status`/`config`/etc. stay in cli.py, so their test monkeypatch
  targets (`qracer.cli._user_dir`, `qracer.cli.load_config`, ...) are unchanged;
  only `test_cli_agents` now imports `_handle_run_agents` from its real home.

cli.py 1520 → 858 lines. Behavior-preserving: 961 passed, pyright 0 errors,
ruff/docs clean, coverage 81%. CLI entrypoint smoke-tested (--help/config/import).
